pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "av-curation-pipeline"
version = "0.1.0"
description = "Multimodal curation pipeline for autonomous-driving datasets (DINOv2 + CLIP + LanceDB + DVC)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Reethika" }]
dependencies = [
    "numpy>=1.26",
    "pyarrow>=15",
    "duckdb>=0.10",
    "pillow>=10",
    "pydantic>=2.5",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "tqdm>=4.66",
    "scikit-learn>=1.4",
    "umap-learn>=0.5.5",
    "hdbscan>=0.8.33",
    "requests>=2.31",
]

[project.optional-dependencies]
ml = [
    "torch>=2.2",
    "transformers>=4.40",
    "open-clip-torch>=2.24",
]
vector = ["lancedb>=0.6"]
nuscenes = ["nuscenes-devkit>=1.1.11"]
dvc = ["dvc>=3.48"]
dev = [
    "pytest>=8",
    "ruff>=0.4",
    "ipykernel>=6.29",
]

[tool.setuptools.packages.find]
include = ["pipeline*", "precompute*"]

[tool.ruff]
line-length = 110
target-version = "py310"
